PileupRenderer refactor
